Refactor: Rename home_screen.dart to home_screen.dart and update import paths
