+++ lib/core/router/router.dart
@@ -4,7 +4,7 @@ import 'package:go_router/go_router.dart';
 import 'package:hummingbird/src/views/login/login_screen.dart';
 import 'package:hummingbird/src/views/splash/splash_screen.dart';
 
-import '../../src/views/home/views/home_screen.dart';
+import '../../src/views/home/home_screen.dart';
 import '../../src/views/more/view/more_screen.dart';
 import '../../src/views/social/views/social_screen.dart';
 import '../../src/views/statistics/views/statistics_screen.dart';
