+++ lib/src/views/home/widgets/mxn_sample.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import '../../../../../core/enum/mxnRate.dart';
-import '../../../../../core/widgets/mxnContainer.dart';
+import '../../../../core/enum/mxnRate.dart';
+import '../../../../core/widgets/mxnContainer.dart';
 
 class MxNSample extends StatelessWidget {
   const MxNSample({super.key});
